ivi-id-agent: new mechanism of desktop surface
with new change on weston 11 and later versions, the surface id
can be changed from the invalid to the valid id. If the client get the
surface with the invalid id in surface_created_event, it can make an unexpectation
if the surface id change to valid on server side.

to avoid this, need to set the invalid id to valid before send the
surface_created_event to clients.

Signed-off-by: Tran Ba Khang(MS/EMC31-XC) <Khang.TranBa@vn.bosch.com>