ivi-controller: new mechanism for screenshot
On weston 11, the weston_renderer structure is moved to internal.
The ivi-controller can't directly access to member of that structure.
So the screenshot functional of controller need to change to new mechanism.

In the new mechanism, the serve side of controller moves to use the
weston_screenshooter_shoot. This api supports to capture the screen.
Client side needs to prepare a shm buffer, then send the buffer to
server side when it's requesting a screen shoot.

This mechanism requires to change the ivi_screenshot interface, screenshoot
request of ivi_wm_screen interface, and surface_screenshoot of ivi_wm interface.

Signed-off-by: Tran Ba Khang(MS/EMC31-XC) <Khang.TranBa@vn.bosch.com>