ivi-id-agent: new mechanism of desktop surface
On weston 11, when xdg surface creating, the ivi-shell is going to emit the
surface_created signal. On older versions, the surface_created signal
is only emited when ivi-shell receiving the surface_configured.

With new logic, the sever sends the invalid id to client, then when receive
the surface_configured, it set new valid id for surface. But client doesn't
know the new valid id of surface, then it makes some requests with invalid id,
the server can't find the right surface.

The new mechanism will set the valid id to valid before send the
surface_created_event to clients.

Signed-off-by: Tran Ba Khang(MS/EMC31-XC) <Khang.TranBa@vn.bosch.com>